Move Swatch component into its own dir
