+++ src/index.js
@@ -40,7 +40,7 @@ export { default as DialogActions } from './dialog/DialogActions'
 export { default as DialogContent } from './dialog/DialogContent'
 export { default as DialogTitle } from './dialog/DialogTitle'
 export { default as Dropdown } from './Dropdown'
-export { default as Swatch } from './colours/Swatch'
+export { default as Swatch } from './Swatch'
 export { default as Switch } from './Switch'
 export { default as TextField } from './TextField'
 export { default as TimePicker } from './pickers/TimePicker'
